requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow>=11.0.0,<12
requests==2.31.0
